a demonstration of pointers in C++
The purpose of this program is to solve a problem using pointer wherever
possible.

A group of students were asked to complete a survey in order to help
improve their school's cafeteria lunches. The survey lists different
menu items and asks whether the student liked or disliked that item. The
purpose of this program is to tally up the results of that survey and
display the results in a table.